refactor: transfer components for migration
