--- src/App.tsx
@@ -1,29 +1,0 @@
-import React from 'react';
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Main from './components/Main';
-import Card from './components/Card';
-import NotFoundPage from './components/NotFoundPage';
-import ErrorBoundary from './components/ErrorBoundary';
-import { ThemeProvider } from './context/ThemeProvider';
-import './App.css';
-
-const App: React.FC = () => {
-  const resetError = () => {};
-
-  return (
-    <ErrorBoundary resetError={resetError}>
-      <ThemeProvider>
-        <Router>
-          <Routes>
-            <Route path="/class-components/" element={<Main searchTerm="" />}>
-              <Route path=":id" element={<Card />} />
-            </Route>
-            <Route path="*" element={<NotFoundPage />} />
-          </Routes>
-        </Router>
-      </ThemeProvider>
-    </ErrorBoundary>
-  );
-};
-
-export default App;
